Legal Doc/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
tensorflow==2.16.1
transformers==4.41.2
pypdf2==3.0.1
python-docx==1.1.2
pdfminer.six==20231228
pymongo==4.6.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1